SAMZA-1050: Make samza-operator independent of avro version
